+++ lib/get.dart
@@ -5,3 +5,4 @@ export 'src/routes.dart';
 export 'src/snack.dart';
 export 'src/bottomsheet.dart';
 export 'src/snack_route.dart';
+export 'src/route_observer.dart';